Fix for cross-realm player name problem
When the client receives the LFG_ROLE_CHECK_ROLE_CHOSEN event, which is
sent whenever a raid member chooses a role, the name of the player is
missing a realm name in all cases, even if the player is from another
realm. This is now taken into account (but makes the mapping from player
name to role choice unavoidably heuristic).